Moved ZeroOneOrMany to util module
