+++ community/cypher/cypher-logical-plans-3.4/src/main/scala/org/neo4j/cypher/internal/v3_4/logical/plans/SeekableArgs.scala
@@ -19,6 +19,7 @@
  */
 package org.neo4j.cypher.internal.v3_4.logical.plans
 
+import org.neo4j.cypher.internal.util.v3_4.{One, ZeroOneOrMany}
 import org.neo4j.cypher.internal.v3_4.expressions._
 
 sealed trait SeekableArgs {
